requirements.txt
asgiref==3.4.1
Django==3.2.7
django-extensions==3.1.5
django-shell-plus==1.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
marshmallow==3.13.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
